Rename to content-style and remove unneeded material
