+++ bin/content-style
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require 'erb_lint'
+require 'content_style'
 require 'YAML'
 
 command = ARGV[0]
@@ -9,13 +9,13 @@ config_file = YAML.load(File.read('config/config.yml'))
 config = config_file.fetch('ContentStyle')
 filetype = config.fetch('filetype')
 dir_path = File.expand_path(command)
-@linter = ERBLint::Linter::ContentStyle.new(config)
+@linter = ContentStyle::Linter.new(config)
 files = Dir.glob(dir_path + '/**/*.' + filetype)
 
 files.each do |file|
   file_content = File.read(file)
   file_path = File.expand_path(file)
-  @violations = @linter.lint_file(ERBLint::Parser.parse(file_content))
+  @violations = @linter.lint_file(ContentStyle::Parser.parse(file_content))
   @violations.each do |v|
     m = v[:message]
     l = v[:line]
