--- lib/erb_lint/linters/content_style.rb
@@ -1,115 +1,0 @@
-# frozen_string_literal: true
-
-require 'nokogiri'
-
-module ERBLint
-  class Linter
-    # Checks for content style guide violations in the text nodes of HTML files.
-    class ContentStyle < Linter
-
-      def initialize(config)
-        @content_ruleset = []
-        config.fetch('rule_set', []).each do |rule|
-          suggestion = rule.fetch('suggestion', '')
-          pattern_description = rule.fetch('pattern_description', '')
-          case_insensitive = rule.fetch('case_insensitive', false)
-          violation_string_or_array = rule.fetch('violation', [])
-          violation_array = [violation_string_or_array].flatten
-          violation_array.each do |violating_pattern|
-            @content_ruleset.push(
-              violating_pattern: violating_pattern,
-              suggestion: suggestion,
-              case_insensitive: case_insensitive,
-              pattern_description: pattern_description
-            )
-          end
-        end
-        @content_ruleset.freeze
-        @addendum = config.fetch('addendum', '')
-      end
-
-      def lint_file(file_tree)
-        errors = []
-        text_nodes = Parser.get_text_nodes(file_tree)
-        text_nodes.each do |text_node|
-          node_has_content = text_node.text =~ /[^\n\s]/
-          node_is_cdata = text_node.cdata?
-          next unless node_has_content && !node_is_cdata
-          content_lines = split_lines(text_node)
-          content_lines.each do |line|
-            errors.concat(generate_errors(line[:text], line[:number]))
-          end
-        end
-        errors
-      end
-
-      private
-
-      def split_lines(text_node)
-        lines = []
-        current_line_number = text_node.parent.line
-        unless text_node.parent.nil?
-          s = StringScanner.new(text_node.text)
-          if s.check_until(/\n/)
-            while (line_content = s.scan_until(/\n/))
-              uri_free_line_content = Parser.strip_uris(line_content)
-              final_line_content = Parser.strip_emails(uri_free_line_content)
-              lines.push(text: final_line_content, number: current_line_number)
-              current_line_number += 1
-            end
-          else
-            line_content = text_node.text
-            uri_free_line_content = Parser.strip_uris(line_content)
-            final_line_content = Parser.strip_emails(uri_free_line_content)
-            lines.push(text: final_line_content, number: current_line_number)
-          end
-        end
-        lines
-      end
-
-      def generate_errors(text, line_number)
-        violated_rules(text).map do |violated_rule|
-          suggestion = violated_rule[:suggestion]
-          pattern_description = violated_rule[:pattern_description]
-          violation = pattern_description.empty? ? violated_rule[:violating_pattern] : pattern_description
-          {
-            line: line_number,
-            message: "Don't use `#{violation}`. Do use `#{suggestion}`.".strip # #{@addendum}".strip
-          }
-        end
-      end
-
-      #TODO: If the line of the error has URIs or email addresses, scan the original line again and match to get the position? OR just do that anyway.
-      #TODO: Should I just find characters here by reversing the order of the scanning? No because there's stuff in the middle of the line that's parsed out. 
-      #So I guess I have to just find that exact violation on that line and then it'll give you the right characters.
-      #TODO: Put the logic back in to check that the violated pattern isn't part of the suggestion
-
-      def violated_rules(text)
-        @content_ruleset.select do |content_rule|
-          violating_pattern = content_rule[:violating_pattern]
-          suggestion = content_rule[:suggestion]
-          xs = 'x' * suggestion.length
-          clean_text = text.gsub(/#{suggestion}/, xs)
-          # Need to write a better test for this
-          case_insensitive = content_rule[:case_insensitive] == true
-          match_case_insensitive = /(#{violating_pattern})\b/i.match(clean_text)
-          match_case_sensitive = /(#{violating_pattern})\b/.match(clean_text)
-          match_ignoring_initial_cap_violations = /[^\.]\s(#{violating_pattern})\b/.match(clean_text)
-          if case_insensitive
-            match_case_insensitive
-          elsif !case_insensitive && suggestion_lowercase_violation_uppercase(suggestion, violating_pattern)
-            match_ignoring_initial_cap_violations
-          else
-            match_case_sensitive
-          end
-        end
-      end
-
-      def suggestion_lowercase_violation_uppercase(suggestion, violating_pattern)
-        suggestion_first_character_lowercase = suggestion.match(/\A[a-z]/)
-        violation_first_character_uppercase = !violating_pattern.match(/\A[A-Z]/)
-        suggestion_first_character_lowercase && violation_first_character_uppercase
-      end
-    end
-  end
-end
