--- spec/erb_lint/fixtures/custom_linter.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module ERBLint
-  class Linter
-    class CustomLinter < Linter
-      include LinterRegistry
-    end
-  end
-end
