--- spec/spec_helper.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require 'erb_lint'
